Issue #146: Fix docker-compose-build.yml usage
We moved the Dockerfile aways from the root directory, so the
docker-compose-build.yml file needs to be adjusted.

I opted for adding a KOHA_IMAGE_OS environment variable to pick the
right Dockerfile. The KOHA_IMAGE one follows the contributed images
naming schema, which is not suitable for reusing here.

To use:
- Run:
  $ KOHA_IMAGE_OS=stretch docker-compose \
                            -f docker-compose-build.yml \
                            -f docker-compose.mariadb9.yml \
                            -p koha up \
                            --force-recreate

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>