Issue #171 - Move git-bz config up in the process
Playing with koha-sandboxes-docker I noticed that the "git bz apply" is sometimes done before the setup is done.

koha-sandboxes-docker wait for 1 min to make sure the container has been built, but it seems that it can be too low.

To prevent sleeping more, we could move the git bz config before the do_all_you_can_do