Issue #176 - Add a KOHA_ELASTICSEARCH flag in the config
Setting KOHA_ELASTICSEARCH to 1 in the config will allow to reindex ES
and set the SearchEngine syspref.

This is related with https://gitlab.com/koha-community/koha-misc4dev/-/merge_requests/43

This should not be the default for now. We should confirm it won't break koha-sandboxes-docker
and wait for 20.05 to be released.