Issue #279 - Prepare selenium tests for installer
Bug 19185 is going to be pushed and we will get selenium test for the
installer and onboarding process.
Those 2 files are a bit different as they require an empty database to
run correctly.
This patch suggests to DROP and CREATE the database to start the
selenium tests.
We cannot run them one after the other, as each require the DB to be
empty.

This is not ideal for several reason:
1. the env vars are duplicated in several places
2. we are dependent on test files in Koha (if we remove or rename them
we will need to adjust here)
3. We don't run everything within a single prove call and so we won't
get the junit file for the first test script

I think we should go as it in a first step, at least see how it goes,
then iterate to make it better if we need.