Issue #285: KTD uses docker-ce
This patch tries to make the README instructions more accurate. KTD
doesn't use the distro's shipped docker.io or docker-compose tools. They
generally just "don't work" so we all actually install the ones provided
by Docker itself.

This instructions are better than the previous ones, but probably still
need a tidy.

Also, Docker Compose V2 is the default for Windows and macOS users, and
it is marked to be GA for Linux (i.e. the default) soon. So we need to
think of phasing out Docker Compose V1 soon, making the install
instructions default to V2, and clean it all. Right now there's a caveat
about V2, it should be the other way around soon. The same for all
sample commands.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>